requirements.txt
Babel==2.12.1
certifi==2022.12.7
click==8.1.3
cloudinary==1.32.0
colorama==0.4.6
Flask==2.2.3
Flask-BabelEx==0.9.4
Flask-Login==0.6.2
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
PyMySQL==1.0.2
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.4
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
